ecommerce/requirements.txt
asgiref==3.8.1
cfgv==3.4.0
distlib==0.3.8
Django==4.2.16
django-cors-headers==4.4.0
djangorestframework==3.15.2
filelock==3.16.0
identify==2.6.0
nodeenv==1.9.1
platformdirs==4.3.2
pre-commit==3.8.0
PyYAML==6.0.2
sqlparse==0.5.1
typing_extensions==4.12.2
virtualenv==20.26.4
